replace Vector2i to Vector2f

+++ cross.h
@@ -6,11 +6,11 @@
 class Cross: Figure
 {
 public:
-  Cross(sf::Vector2i pos, size_t size);
-  explicit Cross(const std::vector< std::pair< sf::Vector2i, sf::Vector2i>> &linesArray);
+  Cross(sf::Vector2f pos, size_t size);
+  explicit Cross(const std::vector< std::pair< sf::Vector2f, sf::Vector2f>> &linesArray);
   ~Cross() override = default;
   void draw(sf::RenderWindow *window);
-  std::vector< std::pair< sf::Vector2i, sf::Vector2i >> linesArray_;
+  std::vector< std::pair< sf::Vector2f, sf::Vector2f >> linesArray_;
 private:
 };
 #endif
