+++ figure.h
@@ -9,14 +9,14 @@ public:
   virtual ~Figure() = default;
   bool isVisible() const;
   void setVisibility(bool status);
-  sf::Vector2i getPos() const;
-  void setPos(sf::Vector2i rhs);
+  sf::Vector2f getPos() const;
+  void setPos(sf::Vector2f rhs);
   size_t getSize() const;
   void setSize(size_t rhs);
 private:
-  std::vector< std::pair< sf::Vector2i, sf::Vector2i >> linesArray_;
+  std::vector< std::pair< sf::Vector2f, sf::Vector2f >> linesArray_;
   bool isVisible_;
-  sf::Vector2i pos_;
+  sf::Vector2f pos_;
   size_t size_;
 };
 #endif
